Fix labels encoding for better compatibility

+++ pyeeglab/dataset/tuh_eeg/tuh_eeg_dataset.py
@@ -26,6 +26,6 @@ class TUHEEGCorpusDataset(Dataset):
             export
         )
         labels = [0 if label == 'normal' else 1 for label in dataset['labels']]
-        labels = np.array(labels).astype('int32').reshape((-1, 1))
+        labels = np.array(labels).astype('float32').reshape((-1, 1))
         dataset = np.array(dataset['data']).astype('float32')
         return dataset, labels
